SCRUM-22: add useSignUp hook
